Global fixes and make compatible with latest version of SayoriOS

+++ Tools/Create RamDisk/SEFS/initrd/src/c/include/env.h
@@ -16,17 +16,17 @@ typedef struct env
   int OS_VERSION_MAJOR;                 ///< Версия ядра
   int OS_VERSION_MINOR;                 ///< Пре-Релиз
   int OS_VERSION_PATCH;                 ///< Патч
-  uint64_t Link_Display;                ///< Ссылка на виртуальный экран
-  uint64_t Link_Time;                   ///< Ссылка на время
+  uint32_t Link_Display;                ///< Ссылка на виртуальный экран
+  uint32_t Link_Time;                   ///< Ссылка на время
   int Display_W;                        ///< Длина экрана
   int Display_H;                        ///< Высота экрана
   int Display_B;                        ///< Смещение экрана
   int Display_P;                        ///< Глубина экрана
   int Display_S;                        ///< Размер буфера
   int Ticks;                            ///< Количество текущих тиков
-  uint64_t RAM_Installed;               ///< Установлено ОЗУ
-  uint64_t RAM_Used;                    ///< Использовано ОЗУ
-  uint64_t RAM_Free;                    ///< Свободно ОЗУ
+  uint32_t RAM_Installed;               ///< Установлено ОЗУ
+  uint32_t RAM_Used;                    ///< Использовано ОЗУ
+  uint32_t RAM_Free;                    ///< Свободно ОЗУ
 } env_t;
 
 struct env* env_io(size_t, size_t);
